sample/custom/deeplearning/requirements.txt
imageio==2.34.1
numpy==1.24.4
pydot==3.0.1
tensorflow==2.16.1
tf_keras==2.16.0
tf_slim==1.1.0
scikit-learn==1.5.0